refactor: MVC 패턴으로 구조 변경
