+++ src/main/java/lotto/controller/LottoFactory.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.controller;
+
+import lotto.domain.Lotto;
 
 @FunctionalInterface
 public interface LottoFactory {
